+++ WebApiCaracterizacion/ModelsMineria/PromediosMinimoOroOR.cs
@@ -9,6 +9,6 @@ namespace WebApiCaracterizacion.ModelsMineria
     {
         public string tipo_plantilla { get; set; }
         public string municipio { get; set; }
-        public decimal promedio { get; set; }
+        public double promedio { get; set; }
     }
 }
